% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{coef,Stocapp-method}
\alias{coef,Stocapp-method}
\title{Method for extracting coefficients from an object in class union "Stocapp"}
\usage{
\S4method{coef}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{coef}}
}
\description{
Method for extracting coefficients from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{coef}
}
