% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{effects,Stocapp-method}
\alias{effects,Stocapp-method}
\title{Method for extracting effects from an object in class union "Stocapp"}
\usage{
\S4method{effects}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{effects}}
}
\description{
Method for extracting effects from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{effects}
}
