% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{fitted,Stocapp-method}
\alias{fitted,Stocapp-method}
\title{Method for extracting fitted values from an object in class union "Stocapp"}
\usage{
\S4method{fitted}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{fitted}}
}
\description{
Method for extracting fitted values from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{fitted.values}
}
