% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{getObject}
\alias{getObject}
\alias{getObject,Stocapp-method}
\title{Accessor to the object in class union "Stocapp"}
\usage{
getObject(x)

\S4method{getObject}{Stocapp}(x)
}
\arguments{
\item{x}{an object of class union "Stocapp"}
}
\description{
Method for obtaining a fitted model within
any object of class union \linkS4class{Stocapp}.
}
\seealso{
\linkS4class{Stocapp}
}
