% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{plot,Stocapp,ANY-method}
\alias{plot,Stocapp,ANY-method}
\title{Method for plotting an object in class union "Stocapp"}
\usage{
\S4method{plot}{Stocapp,ANY}(x, y = NULL, ...)
}
\arguments{
\item{x}{an object of class union "Stocapp"}

\item{y}{not used}

\item{...}{further arguments to pass to \code{plot}}
}
\description{
Method for plotting an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{plot.lm}
}
