% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{predict,Stocapp-method}
\alias{predict,Stocapp-method}
\title{Method for making predictions from an object in class union "Stocapp"}
\usage{
\S4method{predict}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{predict}}
}
\description{
Method for making predictions from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{predict}
}
