% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{residuals,Stocapp-method}
\alias{residuals,Stocapp-method}
\title{Method for extracting residuals from an object in class union "Stocapp"}
\usage{
\S4method{residuals}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{residuals}}
}
\description{
Method for extracting residuals from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{residuals}
}
