% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{show,Stocapp-method}
\alias{show,Stocapp-method}
\title{Method for printing object in class union "Stocapp"}
\usage{
\S4method{show}{Stocapp}(object)
}
\arguments{
\item{object}{an object of class union "Stocapp"}
}
\description{
Method for printing object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}
}
