% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{show,SummaryStocapp-method}
\alias{show,SummaryStocapp-method}
\title{Summarizing a fitted model corrected by the stochastic approximation procedure}
\usage{
\S4method{show}{SummaryStocapp}(object)
}
\arguments{
\item{object}{a summary object of member of \linkS4class{SummaryStocapp}}
}
\description{
Method for printing a \code{summary} of
class union \linkS4class{SummaryStocapp}.
}
\seealso{
\linkS4class{SummaryStocapp}
}
