% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocapp_betareg.R
\name{simulation,betareg-method}
\alias{simulation,betareg-method}
\title{Simulation for a beta regression}
\usage{
\S4method{simulation}{betareg}(object, control = list(...), extra = NULL, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappBetareg}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{stocappControl}}).}

\item{extra}{\code{NULL} by default; extra parameters to pass to simulation.}

\item{...}{further arguments}
}
\description{
simulation method for class \linkS4class{StocappBetareg}
}
