% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocapp_glm.R
\name{simulation,glm-method}
\alias{simulation,glm-method}
\title{Simulation for a Generalized Linear Model regression}
\usage{
\S4method{simulation}{glm}(object, control = list(...), extra = NULL, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappGlm}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{stocappControl}}).}

\item{extra}{\code{NULL} by default; extra parameters to pass to simulation.}

\item{...}{further arguments}
}
\description{
simulation method for class \linkS4class{StocappGlm}
}
