% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocapp_lmer.R
\name{simulation,lmerMod-method}
\alias{simulation,lmerMod-method}
\title{Simulation for linear mixed model regression}
\usage{
\S4method{simulation}{lmerMod}(object, control = list(...), ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappLmer}}

\item{control}{a \code{list} of parameters for controlling the iterative procedure
(see \code{\link{stocappControl}}).}

\item{...}{further arguments.}
}
\description{
simulation method for class \linkS4class{StocappLmer}
}
