% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{Stocapp-class}
\alias{Stocapp-class}
\title{An S4 class union for \code{stocapp}}
\value{
Each member of the union has a \code{slot} with the initial object
corrected by the \code{stocapp} (see \code{\link{getObject}}) and a second \code{slot} with
extra meta data from \code{stocapp} (see \code{\link{getExtra}}).
}
\description{
Members of the union are \linkS4class{StocappBetareg}, \linkS4class{StocappGlm}, \linkS4class{StocappGlmrob2},
\linkS4class{StocappLm}, \linkS4class{StocappLmer}, \linkS4class{StocappNegbin},
\linkS4class{StocappNls}, \linkS4class{StocappVglm}
}
\details{
The `Functions` section describes members of the class union.
}
\seealso{
\code{\link{getExtra}}, \code{\link{getObject}}
}
\author{
Samuel Orso
}
