% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappBetareg-method}
\alias{summary,StocappBetareg-method}
\title{Summarizing a beta regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{StocappBetareg}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappBetareg}}

\item{...}{further arguments passed to \code{summary.betareg}}
}
\description{
summary method for class \linkS4class{StocappBetareg}
}
\seealso{
\link[betareg]{summary.betareg}
}
