% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappGlm-method}
\alias{summary,StocappGlm-method}
\title{Summarizing a Generalized Linear Model regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{StocappGlm}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappGlm}}

\item{...}{further arguments passed to \code{summary.glm}}
}
\description{
summary method for class \linkS4class{StocappGlm}
}
\seealso{
\link[stats]{summary.glm}
}
