% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappLm-method}
\alias{summary,StocappLm-method}
\title{Summarizing a linear regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{StocappLm}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappLm}}

\item{...}{further arguments passed to \code{summary.lm}}
}
\description{
summary method for class \linkS4class{StocappLm}
}
\seealso{
\link[stats]{summary.lm}
}
