% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappNegbin-method}
\alias{summary,StocappNegbin-method}
\title{Summarizing a negative binomial regression fits corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{StocappNegbin}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappNegbin}}

\item{...}{further arguments passed to \code{summary.negbin}}
}
\description{
summary method for class \linkS4class{StocappNegbin}
}
\seealso{
\link[MASS]{summary.negbin}
}
