% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappNls-method}
\alias{summary,StocappNls-method}
\title{Summarizing a nonlinear regression fit corrected by
the iterative bootstrap}
\usage{
\S4method{summary}{StocappNls}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappNls}}

\item{...}{further arguments passed to \code{summary.nls} of \pkg{stats}}
}
\description{
summary method for class \linkS4class{StocappNls}
}
