% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary,StocappVglm-method}
\alias{summary,StocappVglm-method}
\title{Summarizing a vector generalized linear model regression
fit corrected by the iterative bootstrap}
\usage{
\S4method{summary}{StocappVglm}(object, ...)
}
\arguments{
\item{object}{an object of class \linkS4class{StocappVglm}}

\item{...}{further arguments passed to \code{summary.merMod} of \pkg{VGAM}}
}
\description{
summary method for class \linkS4class{StocappVglm}
}
