% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{SummaryStocapp-class}
\alias{SummaryStocapp-class}
\title{An S4 class union for \code{summary}}
\description{
Members of the union are \linkS4class{SummaryStocappBetareg}, \linkS4class{SummaryStocappGlm}, \linkS4class{SummaryStocappLm},
\linkS4class{SummaryStocappLmer}, \linkS4class{SummaryStocappNegbin}, \linkS4class{SummaryStocappNls},
\linkS4class{SummaryStocappVglm}
iterative bootstrap procedure
}
\details{
The `Functions` section describes members of the class union.
}
\author{
Samuel Orso
}
