% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGeneric.R
\name{vcov,Stocapp-method}
\alias{vcov,Stocapp-method}
\title{Method for calculating covariance matrix from an object in class union "Stocapp"}
\usage{
\S4method{vcov}{Stocapp}(object, ...)
}
\arguments{
\item{object}{an object of class union "Stocapp"}

\item{...}{further arguments to pass to \code{vcov}}
}
\description{
Method for calculating covariance matrix from an object in class union "Stocapp"
}
\seealso{
\linkS4class{Stocapp}, \link[stats]{vcov}
}
